+++ src/main/java/org/stellar/sdk/responses/effects/EffectResponse.java
@@ -2,11 +2,12 @@ package org.stellar.sdk.responses.effects;
 
 import com.google.common.base.Optional;
 import com.google.gson.annotations.SerializedName;
-
 import org.stellar.sdk.responses.Link;
 import org.stellar.sdk.responses.MuxedAccount;
 import org.stellar.sdk.responses.Pageable;
 import org.stellar.sdk.responses.Response;
+
+import java.math.BigInteger;
 
 /**
  * Abstract class for effect responses.
@@ -22,7 +23,7 @@ public abstract class EffectResponse extends Response implements Pageable {
   @SerializedName("account_muxed")
   private String accountMuxed;
   @SerializedName("account_muxed_id")
-  private Long accountMuxedId;
+  private BigInteger accountMuxedId;
   @SerializedName("type")
   private String type;
   @SerializedName("created_at")
